Redirect to current URL after authentication (#12)
This PR updates the way that redirects are generated to include the
current request URL as the `redirect-uri` query param of
`_tesseral_next/redirect-login`. This param is then persisted through
the login flow and popped back off when needed. In the case of dev-mode
callbacks, we also include the current request in this URL and the
dev-mode callback first checks for the plain `redirect-uri` param before
failing over to the default param.

The end result is that users are redirected to the page they were
attempting to visit after a successful login.