Preserve query parameters on dev-mode-callback (#15)
This PR fixes a bug in the dev-mode-callback page. Today, that page does
not restore URL search parameters, only URL paths. This PR includes
search parameters as part of the set of data we restore.

For completeness' sake, note that the URL fragment is not possible to
restore, because redirecting to the login page happens primarily from
the server, where fragments are not known to the server.